Update Create Fail Events cnts
This new script changes how failure events are created and counted in the simulation_matrix table. Based on the age of each site, the probability of a failure in a given year goes up so that over time, both the probabiliyt of events and the number of events increases over time.